fix: 🐛 完成了虚拟分页已可用
kernel地址空间已可用

+++ src/xv_r_kernel/trap.rs
@@ -167,7 +167,7 @@ pub fn trap_handler(cx: &mut TrapContext) -> &mut TrapContext {
 }
 fn set_user_trap_entry() {
     unsafe {
-        stvec::write(TRAMPOLINE as usize, TrapMode::Direct);
+        stvec::write(TRAMPOLINE, TrapMode::Direct);
     }
 }
 #[no_mangle]
